adjust scroll up button to only slide out 'scrollup' text on non touch devices
put logic into a custom hook
